Move dot conversion code into separate module
