+++ ibis/src/lib.rs
@@ -13,6 +13,8 @@ mod solution_id;
 mod util;
 #[cfg(feature = "dot")]
 pub mod dot;
+#[cfg(feature = "dot")]
+pub mod recipe_to_dot;
 pub mod recipies;
 extern crate ibis_macros;
